Add "Rate My Quiz" view at the end of protective behavior quiz
This separates out the protective behaviors models into the
"protectivebehaviors" app.

Still TODO is add the custom text based on the
ProtectiveBehaviorsResults.
