+++ requirements.txt
@@ -61,7 +61,7 @@ django-registration==1.1
 django-treebeard==2.0
 django-pagetree==1.0.5
 django-pageblocks==1.0.0
-django-quizblock==1.0.2
+django-quizblock==1.0.3
 django-markwhat==2014.9.20
 gunicorn==19.1.1
 django-infranil==0.1.0
