+++ worth2/templates/protectivebehaviors/protective_behaviors_results.html
@@ -1,10 +1,10 @@
 {% load quizsummary %}
 
-{% with cls="protective-behavior" %}
+{% with cls="protective-behaviors" %}
 
 {% get_quiz_summary user cls as results %}
 
-<div class="protective-behavior-results">
+<div class="protective-behaviors-results">
     <div class="well">
         <h4>Stuff I Do</h4>
         <ul>
