+++ worth2/templates/protectivebehaviors/question.html
@@ -1,0 +1,43 @@
+{% load markup %}
+{% load getresponse %}
+
+{% getquestionresponse question as response %}
+
+<div class="cases">
+  {{question.intro_text|markdown}}
+
+  <div class="casetitle"> Question {{question.display_number}}:</div>
+
+  <div class="casecontent">
+    <div class="casequestion">
+      {{question.text|markdown}}
+      {% if question.answerable %}
+      {% if block.rhetorical %}
+      <form id="form-{{question.id}}" method="post">{% csrf_token %}
+      {% endif %}
+
+      {% include 'protectivebehaviors/rate_my_risk_quiz.html' %}
+
+        {% if block.rhetorical %}
+          </form>
+        {% endif %}
+      {% else %}
+        {% if question.is_short_text %}
+          {% include 'quizblock/short_text.html' %}
+        {% endif %}
+
+        {% if question.is_long_text %}
+          {% include 'quizblock/long_text.html' %}
+        {% endif %}
+      {% endif %}
+    </div>
+
+    {% if block.rhetorical %}
+      {% include 'quizblock/show_answer.html' %}
+    {% else %}
+      {% if response %}
+        {% include 'quizblock/show_answer.html' %}
+      {% endif %}
+  {% endif %}
+  </div>
+</div>
