+++ worth2/templates/protectivebehaviors/quizblock.html
@@ -1,0 +1,18 @@
+{% load markup %}
+{% load getresponse %}
+{{ block.description|markdown }}
+
+{% if request.user.is_anonymous %}
+  <p class="warning">You must be logged in to view this content</p>
+{% else %}
+
+{% for question in block.question_set.all %}
+  {% include 'protectivebehaviors/question.html' %}
+{% endfor %}
+
+{% if block.rhetorical %}
+  <script src="http://mt.ccnmtl.columbia.edu/pedialabs/js/quizshow.js"></script>
+{% endif %}
+
+
+{% endif %}
