+++ worth2/templates/protectivebehaviors/rate_my_risk_quiz.html
@@ -1,0 +1,29 @@
+{% load markup %}
+
+<div class="answers-container">
+    {% for answer in question.answer_set.all %}
+    <div class="answer {{answer.css_extra}}">
+        {% if response %}
+        {% ifequal response.value answer.value %}
+        <span class="yours {% if response.is_correct %}correct{% else %}incorrect{% endif %}"
+              >{{answer.label}}</span>
+        {% else %}
+        {{answer.label}}
+        {% endifequal %}
+        {% else %}
+        <div class="colored-box">
+            <label for="pageblock-{{block.pageblock.id}}-question{{question.id}}-{{answer.value}}">
+                {{answer.label}}
+            </label>
+        </div>
+
+        <div class="clear-box">
+            <input name="pageblock-{{block.pageblock.id}}-question{{question.id}}"
+                   id="pageblock-{{block.pageblock.id}}-question{{question.id}}-{{answer.value}}"
+                   value="{{answer.value}}" type="radio">
+        </div>
+        {% endif %}
+    </div>
+
+    {% endfor %}
+</div>
